pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "AgenticBlocks.IO"
version = "0.8.12"
description = "A building block library for composed agent workflows"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "pydantic>=2.0.0",
    "networkx>=3.0",
    "tenacity>=8.0",
    "litellm>=1.0.0",
    "mcp>=1.0.0",
    "rlms>=0.1.1"
]

[tool.setuptools.packages.find]
where = ["src"]
